+++ test/authorization_test.rb
@@ -16,6 +16,17 @@ class AuthorizationTest < Minitest::Test
     end
   end
 
+  class FakeAllowedAccessibleAuthorizer
+    class << self
+      attr_accessor :last_payload
+
+      def allowed?(**payload)
+        self.last_payload = payload
+        true
+      end
+    end
+  end
+
   class FakeCurrent
     class << self
       attr_accessor :actor, :impersonator
@@ -91,25 +102,72 @@ class AuthorizationTest < Minitest::Test
     end
   end
 
-  def test_authorized_falls_back_to_recording_studio_access_check_when_adapter_is_absent
+  def test_custom_authorization_resolver_can_abstain_to_accessible_authorizer
+    RecordingStudioTrashable.configure do |config|
+      config.use_recording_studio_accessible = true
+      config.authorization_roles = { trash: :edit }
+      config.authorization_resolver = ->(**) {}
+    end
+    FakeAccessibleAuthorizer.last_payload = nil
+
+    RecordingStudioTrashable.const_set(:RecordingStudioAccessible, FakeAccessibleAuthorizer)
+
+    assert RecordingStudioTrashable.authorized?(action: :trash, actor: :user, recording: :recording)
+    assert_equal(
+      { actor: :user, recording: :recording, role: :edit },
+      FakeAccessibleAuthorizer.last_payload
+    )
+  ensure
+    if RecordingStudioTrashable.const_defined?(:RecordingStudioAccessible, false)
+      RecordingStudioTrashable.send(:remove_const, :RecordingStudioAccessible)
+    end
+  end
+
+  def test_authorized_uses_recording_studio_accessible_allowed_adapter
+    RecordingStudioTrashable.configure do |config|
+      config.use_recording_studio_accessible = true
+      config.authorization_roles = { trash: :edit }
+    end
+    FakeAllowedAccessibleAuthorizer.last_payload = nil
+
+    RecordingStudioTrashable.const_set(:RecordingStudioAccessible, FakeAllowedAccessibleAuthorizer)
+
+    assert RecordingStudioTrashable.authorized?(action: :trash, actor: :user, recording: :recording)
+    assert_equal(
+      { actor: :user, recording: :recording, role: :edit },
+      FakeAllowedAccessibleAuthorizer.last_payload
+    )
+  ensure
+    if RecordingStudioTrashable.const_defined?(:RecordingStudioAccessible, false)
+      RecordingStudioTrashable.send(:remove_const, :RecordingStudioAccessible)
+    end
+  end
+
+  def test_authorized_does_not_fall_back_to_removed_core_access_check
     RecordingStudioTrashable.configure do |config|
       config.use_recording_studio_accessible = true
       config.authorization_roles = { trash: :edit }
     end
 
-    payload = nil
+    called = false
+    services_was_defined = RecordingStudio.const_defined?(:Services, false)
+    original_services = RecordingStudio.const_get(:Services) if services_was_defined
+    RecordingStudio.send(:remove_const, :Services) if services_was_defined
+    RecordingStudio.const_set(:Services, Module.new)
+    RecordingStudio::Services.const_set(:AccessCheck, Class.new do
+      define_singleton_method(:allowed?) do |**|
+        called = true
+        true
+      end
+    end)
 
-    RecordingStudio::Services::AccessCheck.stub(:allowed?, lambda { |**kwargs|
-      payload = kwargs
-      true
-    }) do
-      assert RecordingStudioTrashable.authorized?(action: :trash, actor: :user, recording: :recording)
-    end
-
-    assert_equal(
-      { actor: :user, recording: :recording, role: :edit },
-      payload
-    )
+    refute RecordingStudioTrashable.authorized?(action: :trash, actor: :user, recording: :recording)
+    refute called
+    refute_includes File.read(File.expand_path("../lib/recording_studio_trashable/authorization.rb", __dir__)),
+                    "AccessCheck"
+  ensure
+    RecordingStudio.send(:remove_const, :Services) if RecordingStudio.const_defined?(:Services, false)
+    RecordingStudio.const_set(:Services, original_services) if services_was_defined
   end
 
   def test_current_actor_and_impersonator_fall_back_to_current_attributes
@@ -146,7 +204,7 @@ class AuthorizationTest < Minitest::Test
     )
     assert_includes(
       source,
-      "success_message: -> { \"\#{recording_studio_trashable_recording_label(@recording)} permantly deleted\" }"
+      "success_message: -> { \"\#{recording_studio_trashable_recording_label(@recording)} permanently deleted\" }"
     )
     assert_includes(
       source,