+++ test/dummy/app/models/folder.rb
@@ -1,2 +1,4 @@
 class Folder < ApplicationRecord
+  recording_studio_recordable label: "Folder", plural_label: "Folders", root: false,
+                              allowed_parent_types: %w[Workspace Project Folder]
 end