+++ test/dummy/app/models/page.rb
@@ -1,3 +1,6 @@
 class Page < ApplicationRecord
+  recording_studio_recordable label: "Page", plural_label: "Pages", root: false,
+                              allowed_parent_types: %w[Workspace Project Folder Page]
+
   include RecordingStudio::Capabilities::Trashable.to
 end